feat: use violations set

+++ shop-mngt-domain/src/main/java/com/example/shop/domain/usecase/user/DeleteUserUsecase.java
@@ -1,32 +1,40 @@
 package com.example.shop.domain.usecase.user;
 
+import com.example.shop.domain.dto.SimpleViolation;
 import com.example.shop.domain.repository.UserRepository;
+import com.example.shop.domain.validators.Validator;
+import com.example.shop.domain.validators.user.UserExistenceValidator;
+import com.example.shop.domain.validators.user.UserIdValidator;
 
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
 import java.util.UUID;
 
 public class DeleteUserUsecase {
 
     private final UserRepository userRepository;
+    private final List<Validator<UUID>> validators;
 
     public DeleteUserUsecase(UserRepository userRepository) {
         this.userRepository = userRepository;
+        this.validators = List.of(
+                new UserIdValidator(),
+                new UserExistenceValidator(userRepository)
+        );
     }
 
     public void execute(UUID id) {
-        checkIfNull(id);
-        checkIfUserExists(id);
+        Set<SimpleViolation> violations = new HashSet<>();
+
+        for (Validator<UUID> validator : validators) {
+            violations.addAll(validator.validate(id));
+        }
+
+        if (!violations.isEmpty()) {
+            throw new IllegalArgumentException("Validation failed: " + violations);
+        }
 
         userRepository.deleteById(id);
-    }
-    private void checkIfNull(UUID id) {
-        if (id == null) {
-            throw new IllegalArgumentException("User ID cannot be null");
-        }
-    }
-
-    private void checkIfUserExists(UUID id) {
-        if (!userRepository.existsById(id)) {
-            throw new IllegalArgumentException("User with ID " + id + " does not exist");
-        }
     }
 }
